+++ rust/operator-binary/src/hbase_controller.rs
@@ -355,8 +355,6 @@ pub async fn reconcile_hbase(
 
     let client = &ctx.client;
 
-    validate_cr(hbase)?;
-
     let resolved_product_image = hbase
         .spec
         .image
@@ -630,13 +628,8 @@ fn build_rolegroup_config_map(
                 );
             }
             PropertyNameKind::File(file_name) if file_name == HBASE_ENV_SH => {
-                let mut hbase_env_config = build_hbase_env_sh(
-                    hbase,
-                    merged_config,
-                    &hbase_role,
-                    &rolegroup.role_group,
-                    &resolved_product_image.product_version,
-                )?;
+                let mut hbase_env_config =
+                    build_hbase_env_sh(hbase, merged_config, &hbase_role, &rolegroup.role_group)?;
 
                 // configOverride come last
                 hbase_env_config.extend(config.clone());
@@ -720,15 +713,11 @@ fn build_rolegroup_config_map(
         builder.add_data(SSL_CLIENT_XML, ssl_client_xml);
     }
 
-    extend_role_group_config_map(
-        rolegroup,
-        merged_config.logging(),
-        &mut builder,
-        &resolved_product_image.product_version,
-    )
-    .context(InvalidLoggingConfigSnafu {
-        cm_name: rolegroup.object_name(),
-    })?;
+    extend_role_group_config_map(rolegroup, merged_config.logging(), &mut builder).context(
+        InvalidLoggingConfigSnafu {
+            cm_name: rolegroup.object_name(),
+        },
+    )?;
 
     builder.build().map_err(|e| Error::BuildRoleGroupConfig {
         source: e,
@@ -746,7 +735,7 @@ fn build_rolegroup_service(
     resolved_product_image: &ResolvedProductImage,
 ) -> Result<Service> {
     let ports = hbase
-        .ports(hbase_role, &resolved_product_image.product_version)
+        .ports(hbase_role)
         .into_iter()
         .map(|(name, value)| ServicePort {
             name: Some(name),
@@ -811,7 +800,7 @@ fn build_rolegroup_statefulset(
     let hbase_version = &resolved_product_image.app_version_label;
 
     let ports = hbase
-        .ports(hbase_role, &resolved_product_image.product_version)
+        .ports(hbase_role)
         .into_iter()
         .map(|(name, value)| ContainerPort {
             name: Some(name),
@@ -1172,7 +1161,6 @@ fn build_hbase_env_sh(
     merged_config: &AnyServiceConfig,
     hbase_role: &HbaseRole,
     role_group: &str,
-    product_version: &str,
 ) -> Result<BTreeMap<String, String>, Error> {
     let mut result = BTreeMap::new();
 
@@ -1187,7 +1175,7 @@ fn build_hbase_env_sh(
         construct_global_jvm_args(hbase.has_kerberos_enabled()),
     );
     let role_specific_non_heap_jvm_args =
-        construct_role_specific_non_heap_jvm_args(hbase, hbase_role, role_group, product_version)
+        construct_role_specific_non_heap_jvm_args(hbase, hbase_role, role_group)
             .context(ConstructJvmArgumentSnafu)?;
 
     match hbase_role {
@@ -1212,24 +1200,6 @@ fn build_hbase_env_sh(
     }
 
     Ok(result)
-}
-
-/// Ensures that no authorization is configured for HBase versions that do not support it.
-/// In the future, such validations should be moved to the CRD CEL rules which are much more flexible
-/// and have to added benefit that invalid CRs are rejected by the API server.
-/// A requirement for this is that the minimum supported Kubernetes version is 1.29.
-fn validate_cr(hbase: &v1alpha1::HbaseCluster) -> Result<()> {
-    tracing::info!("Begin CR validation");
-
-    let hbase_version = hbase.spec.image.product_version();
-    let authorization = hbase.spec.cluster_config.authorization.is_some();
-
-    if hbase_version.starts_with("2.4") && authorization {
-        tracing::error!("Invalid custom resource");
-        return Err(Error::AuthorizationNotSupported);
-    }
-    tracing::info!("End CR validation");
-    Ok(())
 }
 
 #[cfg(test)]